Add Weighted Move Average
Weighted Move Average as 0.5, 0.25, 0.125, 0.0625, 0.0625

+++ settime.c
@@ -11,7 +11,7 @@
 #include "absacc.h"
 
 uchar Inctime;						//时间按键增加标志
-float Real_Count_Display[8];			//AA1*原为Redrawtime
+double Real_Count_Display[8];			//AA1*原为Redrawtime
 uchar Setted_Time[8];				//AB1+
 uchar Init_Time=10;					//AD1*默认时间为10s
